refactor: hookに力だせていない処理をリファクタ
